requirements.txt
customtkinter
google-generativeai
google-genai
openai-whisper
edge-tts
sounddevice
soundfile
numpy

# # on Ubuntu or Debian
# sudo apt update && sudo apt install ffmpeg

# # on Arch Linux
# sudo pacman -S ffmpeg

# # on MacOS using Homebrew (https://brew.sh/)
# brew install ffmpeg

# # on Windows using Chocolatey (https://chocolatey.org/)
# choco install ffmpeg

# # on Windows using Scoop (https://scoop.sh/)!!
# scoop install ffmpeg